requirements.txt
amqp==5.0.2
arabic-reshaper==2.1.4
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
Brotli==1.0.9
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==36.0.2
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-extensions==3.1.1
django-filter==2.4.0
django-markdown-deux==1.0.5
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-pagedown==2.2.0
djangorestframework==3.12.2
django-semanticui-forms==1.6.5
django-unused-media==0.2.2
future==0.18.2
# gunicorn
html-text==0.5.2
html5lib==1.1
idna==2.10
ipython==8.2.0
ipython-genutils==0.2.0
jsonfield==3.1.0
jsonschema==3.2.0
kombu==5.0.2
lxml==4.8.0
markdown2==2.4.2
numpy==1.22.3
oauthlib==3.1.0
openpyxl==3.0.7
Pillow==8.1.0
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.6.0
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.2
swapper==1.1.2.post1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0